+++ src/gameobjects/sprite/SpriteCanvasRenderer.js
@@ -21,7 +21,7 @@
  */
 var SpriteCanvasRenderer = function (renderer, src, interpolationPercentage, camera, parentMatrix)
 {
-    renderer.drawImage(src, camera, parentMatrix);
+    renderer.batchSprite(src, src.frame, camera, parentMatrix);
 };
 
 module.exports = SpriteCanvasRenderer;