+++ src/gameobjects/text/static/TextCanvasRenderer.js
@@ -21,58 +21,10 @@
  */
 var TextCanvasRenderer = function (renderer, src, interpolationPercentage, camera, parentMatrix)
 {
-    if (src.text === '')
+    if (src.text !== '')
     {
-        return;
+        renderer.batchSprite(src, src.frame, camera, parentMatrix);
     }
-    
-    var ctx = renderer.currentContext;
-
-    var alpha = camera.alpha * src.alpha;
-
-    if (alpha === 0)
-    {
-        //  Nothing to see, so abort early
-        return;
-    }
-    
-    //  Blend Mode
-    ctx.globalCompositeOperation = renderer.blendModes[src.blendMode];
-
-    //  Alpha
-    ctx.globalAlpha = alpha;
-
-    var canvas = src.canvas;
-
-    ctx.save();
-
-    if (parentMatrix)
-    {
-        parentMatrix.copyToContext(ctx);
-    }
-
-    var tx = src.x - camera.scrollX * src.scrollFactorX;
-    var ty = src.y - camera.scrollY * src.scrollFactorY;
-
-    if (camera.roundPixels)
-    {
-        tx |= 0;
-        ty |= 0;
-    }
-
-    ctx.translate(tx, ty);
-
-    ctx.rotate(src.rotation);
-
-    ctx.scale(src.scaleX, src.scaleY);
-
-    ctx.translate(canvas.width * (src.flipX ? 1 : 0), canvas.height * (src.flipY ? 1 : 0));
-
-    ctx.scale(src.flipX ? -1 : 1, src.flipY ? -1 : 1);
-
-    ctx.drawImage(canvas, 0, 0, canvas.width, canvas.height, -src.displayOriginX, -src.displayOriginY, canvas.width / src.style.resolution, canvas.height / src.style.resolution);
-
-    ctx.restore();
 };
 
 module.exports = TextCanvasRenderer;